cargo-bpf: add bindings feature
The bindings feature depends on the minimal set of deps required to run
cargo_bpf::bindgen::*. In their build.rs, probes commonly just need to
generate bindings for kernel/user headers, so this improves compile
times significantly.
